+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:expense_tracker_app/expenses.dart';
+import 'package:expense_tracker_app/widgets/expenses.dart';
 import 'package:flutter/material.dart';
 
 void main() {
